+++ packages/core/src/utils/make-namespaced.ts
@@ -26,9 +26,10 @@ const OUTPUT_SELECTION = {
  *
  * @param input The original solc input.
  * @param output The original solc output.
+ * @param solcVersion The version of the solc compiler that was originally used to compile the input.
  * @returns The modified solc input with storage layout that includes namespaced type information.
  */
-export function makeNamespacedInput(input: SolcInput, output: SolcOutput): SolcInput {
+export function makeNamespacedInput(input: SolcInput, output: SolcOutput, solcVersion?: string): SolcInput {
   const modifiedSources: Record<string, { content?: string }> = {};
 
   for (const [sourcePath] of Object.entries(input.sources)) {
@@ -150,10 +151,66 @@ export function makeNamespacedInput(input: SolcInput, output: SolcOutput): SolcI
       }
     }
 
-    modifiedSources[sourcePath] = { ...source, content: getModifiedSource(orig, modifications) };
+    const modifiedSource = tryRemoveNonStructNatSpec(getModifiedSource(orig, modifications), solcVersion);
+
+    modifiedSources[sourcePath] = { ...source, content: modifiedSource };
   }
 
   return { ...input, sources: modifiedSources, settings: { ...input.settings, outputSelection: OUTPUT_SELECTION } };
+}
+
+/**
+ * If Slang is supported for the current platform and we have the compiler version available,
+ * use Slang to parse and remove all NatSpec comments that do not precede a struct definition and return the modified content.
+ *
+ * Otherwise, return the original content.
+ */
+function tryRemoveNonStructNatSpec(origContent: string, solcVersion: string | undefined): string {
+  const natSpecRemovals: Modification[] = [];
+
+  if (solcVersion !== undefined && tryRequire('@nomicfoundation/slang')) {
+    /* eslint-disable @typescript-eslint/no-var-requires */
+    const { Language } = require('@nomicfoundation/slang/language');
+    const { NonterminalKind, TerminalKind } = require('@nomicfoundation/slang/kinds');
+    const { TerminalNode } = require('@nomicfoundation/slang/cst');
+    const { isTrivia } = require('./slang/trivia');
+    /* eslint-enable @typescript-eslint/no-var-requires */
+
+    const language = new Language(solcVersion);
+    const parseOutput = language.parse(NonterminalKind.SourceUnit, origContent);
+    const cursor = parseOutput.createTreeCursor();
+
+    while (
+      cursor.goToNextTerminalWithKinds([TerminalKind.MultiLineNatSpecComment, TerminalKind.SingleLineNatSpecComment])
+    ) {
+      // Lookahead to determine if the next non-trivia node is the struct keyword.
+      // If so, this NatSpec is part of the struct definition and should not be removed.
+      const lookaheadCursor = cursor.clone();
+      while (lookaheadCursor.goToNextTerminal() && isTrivia(lookaheadCursor.node())) {
+        // skip over trivia nodes
+      }
+      if (lookaheadCursor.node().kind === TerminalKind.StructKeyword) {
+        continue;
+      }
+
+      const triviaNode = cursor.node();
+      assert(triviaNode instanceof TerminalNode);
+      natSpecRemovals.push(makeDeleteRange(cursor.textRange.start.utf8, cursor.textRange.end.utf8));
+    }
+    return getModifiedSource(Buffer.from(origContent, 'utf8'), natSpecRemovals);
+  } else {
+    return origContent;
+  }
+}
+
+function tryRequire(id: string) {
+  try {
+    require(id);
+    return true;
+  } catch (e: any) {
+    // do nothing
+  }
+  return false;
 }
 
 interface Modification {
@@ -238,7 +295,11 @@ function makeDelete(node: Node, orig: Buffer): Modification {
       end += 1;
     }
   }
-  return { start: positions.start, end };
+  return makeDeleteRange(positions.start, end);
+}
+
+function makeDeleteRange(start: number, end: number): Modification {
+  return { start, end };
 }
 
 function getModifiedSource(orig: Buffer, modifications: Modification[]): string {